+++ app/Http/Controllers/Admin/RoleController.php
@@ -2,6 +2,7 @@
 
 namespace App\Http\Controllers\Admin;
 
+use App\Permission;
 use App\Role;
 use Illuminate\Http\Request;
 use App\Http\Controllers\Controller;
@@ -16,10 +17,13 @@ class RoleController extends Controller
      */
     public function index()
     {
+        $this->authorize('view', Role::class);
+
         $roles = Role::all();
 
         return view('admin.roles.index', compact('roles'));
     }
+
     /**
      * Show the form for creating a new group.
      *
@@ -27,7 +31,11 @@ class RoleController extends Controller
      */
     public function create()
     {
-        return view('admin.roles.create');
+        $this->authorize('create', Role::class);
+
+        $permissions = Permission::all();
+
+        return view('admin.roles.create', compact('permissions'));
     }
 
     /**
@@ -38,13 +46,14 @@ class RoleController extends Controller
      */
     public function store(StoreRole $request)
     {
-        $role = new Role($request->only('id', 'name'));
-        // @TODO: manage permissions
-        $role->permissions = [];
+        $this->authorize('create', Role::class);
+
+        $role = new Role($request->only('id', 'name', 'permissions'));
         $role->save();
 
         return redirect()->route('admin.roles.index');
     }
+
     /**
      * Display the specified role.
      *
@@ -53,8 +62,11 @@ class RoleController extends Controller
      */
     public function show(Role $role)
     {
+        $this->authorize('view', $role);
+
         return view('admin.roles.show', compact('role'));
     }
+
     /**
      * Show the form for editing the specified role.
      *
@@ -63,13 +75,17 @@ class RoleController extends Controller
      */
     public function edit(Role $role)
     {
+        $this->authorize('update', $role);
+
         if ($role->isProtected()) {
             flash("Il ruolo {$role->name} non può essere modificato", 'error');
 
             return redirect()->back();
         }
 
-        return view('admin.roles.edit', compact('role'));
+        $permissions = Permission::all();
+
+        return view('admin.roles.edit', compact('role', 'permissions'));
     }
 
     /**
@@ -81,13 +97,15 @@ class RoleController extends Controller
      */
     public function update(Role $role, StoreRole $request)
     {
+        $this->authorize('update', $role);
+
         if ($role->isProtected()) {
             flash("Il ruolo {$role->name} non può essere modificato", 'error');
 
             return redirect()->back();
         }
 
-        $role->update($request->only('name'));
+        $role->update($request->only('name', 'permissions'));
 
         return redirect()->route('admin.roles.index');
     }
@@ -100,7 +118,10 @@ class RoleController extends Controller
      */
     public function destroy(Role $role)
     {
+        $this->authorize('delete', $role);
+
         $role->delete();
 
         return redirect()->route('admin.roles.index');
-    }}
+    }
+}
